fixes for windows build

+++ vef/tilecutter.cpp
@@ -203,7 +203,7 @@ void Cutter::operator()(/**vt::ExternalProgress &progress*/)
     {
         const auto convertor(vef2world());
         UTILITY_OMP(for schedule(dynamic, 1))
-        for (std::size_t i = 0; i < windows_.size(); ++i) {
+        for (std::int64_t i = 0; i < static_cast<std::int64_t>(windows_.size()); ++i) {
             windowCut(windows_[i], convertor);
         }
     }
